requirements.txt
python-telegram-bot>=21.0
python-dotenv==1.0.0
requests==2.31.0
maigret
yt-dlp

#source venv/bin/activate && TELEGRAM_BOT_TOKEN="8640746951:AAHQO0stCvKdRP2wr4WalT4N4SegRmwTO_0" python bot.py